deps/v8/test/inspector/regress/regress-crbug-1197392-expected.txt

Regression test for crbug.com/1197392

Running test: testInvalidSamplingInterval
{
    error : {
        code : -32000
        message : Invalid sampling interval
    }
    id : <messageId>
}
